moved features to separate file
